backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
phonenumberslite==8.13.13
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
